Make max recent files a constant rather than an enum and use it for defining the size of the array as well.

+++ depthmapX/mainwindow.cpp
@@ -2034,7 +2034,7 @@ void MainWindow::updateRecentFileActions()
 
     QStringList files = settings.value("recentFileList").toStringList();
 
-    int numRecentFiles = qMin(files.size(), (int)MaxRecentFiles);
+    int numRecentFiles = qMin(files.size(), MaxRecentFiles);
 
     for (int i = 0; i < numRecentFiles; ++i) {
         QString text = tr("&%1 %2").arg(i + 1).arg(strippedName(files[i]));
